Huge updates. A ton of new documentation. Added shader-based Gabor class. Gabor stuff moved to own files. Keyboard changed to more fully support new GLFW keycode information. Unicode and scancode are both included with CX_Keyboard::Event now. The default keycode type has changed to the GLFW code, because that does not change with modifier keys. Letter keys are not lowercase anymore. Mouse now tracks cursor position internally instead of using oF functions and as a result now inverts properly when inverting the y-axis. CX_SlidePresenter now has a mechanism to allow user code to check what slide was the last slide to be presented (getLastPresentedSlideName()). Also some improvements in the presentation information printout. CX_Display now has functions for inverting the y-axis. Also, makeFbo to make a framebuffer with the same dimensions and settings as the main display.

+++ src/CX_Clock.cpp
@@ -36,7 +36,6 @@ std::string CX_Clock::precisionTest (unsigned int iterations) {
 		durations[i] = t2 - t1;
 	}
 
-	//uint64_t differenceSum = 0;
 	long long maxDifference = 0;
 	long long minDifference = std::numeric_limits<long long>::max();
 	long long minNonzeroDuration = std::numeric_limits<long long>::max();
@@ -45,8 +44,6 @@ std::string CX_Clock::precisionTest (unsigned int iterations) {
 
 		long long duration = durations[i];
 		durations[i] = duration;
-
-		//differenceSum += duration;
 
 		if (duration > maxDifference) {
 			maxDifference = duration;
@@ -63,7 +60,7 @@ std::string CX_Clock::precisionTest (unsigned int iterations) {
 
 	if (minNonzeroDuration > 1000000) {
 		CX::Instances::Log.warning("CX_Clock") << "The precision of the system clock used by CX_Clock appears to be worse than "
-			"millisecond precision. Observed tick period of the system clock is " << minNonzeroDuration / 1000000.0 << " milliseconds. "
+			"millisecond precision. Observed tick period of the system clock is " << (double)minNonzeroDuration / 1000000.0 << " milliseconds. "
 			"See CX_Clock::setImplementation() for information about using a different underlying clock implementation.";
 	}
